backend-end!
Go front!

+++ src/controllers/videoControllers.js
@@ -1,5 +1,5 @@
 import Video from "../models/Video"; //model 수입하기.
-//import {formatHashtags} from "../models/Video"; //함수를 사용한 경우
+import User from "../models/User"; //model 수입하기.
 
 //Home
 
@@ -28,9 +28,13 @@ export const search = async(req, res) => {
 
 export const getEdit = async(req, res) =>{
     const { id } = req.params;
+    const {user : {_id}} = req.session;
     const video = await Video.findById(id);
     if(!video){
         return res.status(404).render("404", { pageTitle : "Video not found."})
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
     }
     return res.render("edit", { pageTitle : `Editing :${video.title}`, video});
 };
@@ -42,16 +46,17 @@ export const postEdit = async (req, res) =>{
         return res.status(404).render("404", { pageTitle : "Video not found."})
     }
     Video.findByIdAndUpdate(id, {
-
         title,
         description,
         hashtags : Video.formatHashtags(hashtags),
     });
     return res.redirect(`/videos/${id}`);
 };
+
 export const watch = async (req, res) => {
-    const { id } = req.params; // == const id = req.params.id
-    const video = await Video.findById(id);
+    const { id } = req.params; 
+    const video = await Video.findById(id).populate("owner");
+    console.log(video)
     if(video === null){
         return res.render("404", {pageTitle : "Video not found."})
     }
@@ -63,14 +68,20 @@ export const getUpload = (req, res) => {
     return res.render("upload", {pageTitle : "Upload Video"});
 };
 export const postUpload = async (req, res) => {
+    const {user : {_id}} = req.session;
+    const fileUrl = req.file.path
     const {title, description, hashtags} = req.body;
     try {
-        await Video.create({ //model에는 타입만 지정했지만, 여기서는 직접 값을 넣어준다. 
-            title, //title : title 과 같음. 여기서 후자는 위의 title의 값을 뜻함.
+        const newVideo = await Video.create({ 
+            fileUrl,
+            title,
+            owner : _id,
             description,
             hashtags : Video.formatHashtags(hashtags)
         });
-        console.log(Video);
+        const user = await User.findById(_id);
+        user.videos.push(newVideo.id);
+        user.save(); //유저의 정보를 저장함. 위에서 고쳤기 때문에 사용해줘야 update된 내용이 들어간다.
         return res.redirect("/");
     }catch(error){
         console.log(error.message);
@@ -80,6 +91,14 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
+    const {user : {_id}} = req.session;
+    const video = await Video.findById(id);
+    if(!video){
+        return res.status(404).render("404", {pageTitle : "Video not found. "})
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 };
